Still trying to improve the performance. Local tests OK.

+++ priv/repo/migrations/20250709000325_create_payments.exs
@@ -4,7 +4,7 @@ defmodule Rinha.Repo.Migrations.CreatePayments do
   def change do
     create table(:payments) do
       add :correlation_id, :string, null: false
-      add :amount, :integer, null: false
+      add :amount, :float, null: false
       add :processor, :string
       add :requested_at, :utc_datetime_usec
     end
